fixed broken imports in alignment module moved template library code back to corridor
