+++ freecad/trails/corridor/template/SketchTemplate.py
