+++ freecad/trails/corridor/template/ViewTemplateLibrary.py
